+++ resources/db/test-db/migrations/deploy/create-twitter-users-table.sql
@@ -3,7 +3,7 @@
 BEGIN;
 
 CREATE TABLE twitter.users (
-  id integer PRIMARY KEY,
+  id bigint PRIMARY KEY,
   screen_name text NOT NULL,
   name text NOT NULL,
   description text,
